skeleton for implementing new connects
